spark_job/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
kafka-python==2.2.9
    # via 09-spark-redpanda (pyproject.toml)
py4j==0.10.9.7
    # via pyspark
pyspark==3.5.5
    # via 09-spark-redpanda (pyproject.toml)
python-dotenv==1.1.0
    # via 09-spark-redpanda (pyproject.toml)
